# docs/api/openapi.json
{
...
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Catalogo de juegos de mesa buscables por materiales, jugadores y reglas.",
    "title": "QueJugamos API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/games/{slug}/comments": {
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "default": "<string>",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "default": "<string>",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "bodyMd": {
                    "default": "<string>",
                    "description": "Comment body in Markdown.",
                    "example": "Works well with kids if you shorten the rounds.",
                    "type": "string"
                  },
                  "parentId": {
                    "default": "<uuid>",
                    "description": "Parent comment UUID when replying to another comment.",
                    "example": "22222222-2222-2222-2222-222222222222",
                    "format": "uuid",
                    "type": "string"
                  },
                  "userId": {
                    "default": "<uuid>",
                    "description": "UUID of the comment author.",
                    "example": "11111111-1111-1111-1111-111111111111",
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "bodyMd"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
...
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
...
}